Sessions-CheckUser-Secured

+++ classes/Ijdb/Controllers/Joke.php
@@ -60,13 +60,31 @@ class Joke
     }
     public function delete()
     {
-        $this->jokesTable->delete($_POST['id']);
+        $author = $this->authentication->getUser();
 
+        $joke = $this->jokesTable->findById($_POST['id']);
+        
+        if ($joke['authorId'] != $author['id']) {
+            return;
+        }
+        
+        $this->jokesTable->delete($_POST['id']);
+        
+        
         header('location: /joke/public/joke/list');
     }
     public function saveEdit()
     {
         $author = $this->authentication->getUser();
+
+        if (isset($_GET['id'])) {
+            $joke = $this->jokesTable->findById($_GET['id']);
+
+            if ($joke['authorId'] != $author['id']) {
+                return;
+            }
+        }
+
         $joke = $_POST['joke'];
         $joke['jokedate'] = new \DateTime();
         $joke['authorId'] = $author['id'];
@@ -75,16 +93,20 @@ class Joke
     }
     public function edit()
     {
+        $author = $this->authentication->getUser();
+
         if (isset($_GET['id'])) {
             $joke = $this->jokesTable->findById($_GET['id']);
         }
 
         $title = 'Edit joke';
+
         return [
             'template' => 'editjoke.html.php',
             'title' => $title,
             'variables' => [
-                'joke' => $joke ?? null
+                'joke' => $joke ?? null,
+                'userId' => $author['id'] ?? null
             ]
         ];
     }
